Add team color visualization tool, improve index pages
Add the tool used to visualize the team color change of ddnet/ddnet#7129.

Improve the layout of the index page and also add the team color tool to it.

Make the layout of Map and Demo Viewer index page consistent with this. Remove unnecessary duplicate specification of the UTF-8 encoding. Add viewport specification.